refactor: move StringIsUppercase to validate package
