+++ internal/service/advancedcluster/model_advanced_cluster_test.go
@@ -120,39 +120,6 @@ func TestUpgradeRefreshFunc(t *testing.T) {
 	}
 }
 
-func TestStringIsUppercase(t *testing.T) {
-	testCases := []struct {
-		name          string
-		expectedError bool
-	}{
-		{
-			name:          "AWS",
-			expectedError: false,
-		},
-		{
-			name:          "aws",
-			expectedError: true,
-		},
-		{
-			name:          "",
-			expectedError: false,
-		},
-		{
-			name:          "AwS",
-			expectedError: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			diag := advancedcluster.StringIsUppercase()(tc.name, nil)
-			if diag.HasError() != tc.expectedError {
-				t.Errorf("Case %s: Received unexpected error: %v", tc.name, diag[0].Summary)
-			}
-		})
-	}
-}
-
 func TestResourceListAdvancedRefreshFunc(t *testing.T) {
 	testCases := []struct {
 		mockCluster    *admin.PaginatedAdvancedClusterDescription
